% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-result.R
\name{AnalysisResult-accessors}
\alias{AnalysisResult-accessors}
\alias{resultType}
\alias{resultType,AnalysisResult-method}
\alias{resultValue}
\alias{resultValue,AnalysisResult-method}
\alias{resultValue,PhysioBiomarker-method}
\alias{biomarkerValue}
\alias{biomarkerValue,PhysioBiomarker-method}
\title{Accessors for analysis results}
\usage{
resultType(x)

\S4method{resultType}{AnalysisResult}(x)

resultValue(x)

\S4method{resultValue}{AnalysisResult}(x)

\S4method{resultValue}{PhysioBiomarker}(x)

biomarkerValue(x)

\S4method{biomarkerValue}{PhysioBiomarker}(x)
}
\arguments{
\item{x}{An \code{AnalysisResult} or \code{PhysioBiomarker}.}
}
\value{
\code{resultType()} the type tag; \code{resultValue()} the payload
(or scalar value for a biomarker); \code{biomarkerValue()} the scalar value.
}
\description{
Accessors for analysis results
}
\examples{
res <- AnalysisResult("hrv_time", result = list(sdnn = 42, rmssd = 30))
resultType(res)
resultValue(res)

bm <- PhysioBiomarker("SDNN", 42, unit = "ms")
biomarkerValue(bm)
}
