% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis-result.R
\name{AnalysisResult-carrier}
\alias{AnalysisResult-carrier}
\alias{estimateOf}
\alias{estimateOf,AnalysisResult-method}
\alias{uncertaintyOf}
\alias{uncertaintyOf,AnalysisResult-method}
\alias{provenanceOf}
\alias{provenanceOf,AnalysisResult-method}
\alias{estimandOf}
\alias{estimandOf,AnalysisResult-method}
\title{Estimand / uncertainty carrier accessors}
\usage{
estimateOf(x)

\S4method{estimateOf}{AnalysisResult}(x)

uncertaintyOf(x)

\S4method{uncertaintyOf}{AnalysisResult}(x)

provenanceOf(x)

\S4method{provenanceOf}{AnalysisResult}(x)

estimandOf(x)

\S4method{estimandOf}{AnalysisResult}(x)
}
\arguments{
\item{x}{An \code{AnalysisResult}.}
}
\value{
\code{estimateOf()} the point estimate; \code{uncertaintyOf()} the
interval list; \code{provenanceOf()} the provenance; \code{estimandOf()} the
ICH E9(R1) estimand list.
}
\description{
Estimand / uncertainty carrier accessors
}
\examples{
r <- AnalysisResult("rom", estimate = 118,
  uncertainty = list(type = "conformal", level = 0.9, lower = 104,
                     upper = 132))
estimateOf(r)
uncertaintyOf(r)
}
