% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-epoch.R
\name{averageEpochs}
\alias{averageEpochs}
\title{Average epochs}
\usage{
averageEpochs(x, by = NULL)
}
\arguments{
\item{x}{An epoched PhysioExperiment object.}

\item{by}{Optional column name in epoch_info to group by.}
}
\value{
A PhysioExperiment object with averaged epochs.
}
\description{
Computes the average across epochs, optionally by condition.
}
