% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui-launcher.R
\name{checkGUIDependencies}
\alias{checkGUIDependencies}
\title{Check GUI Dependencies}
\usage{
checkGUIDependencies()
}
\value{
A list with the status of each dependency.
}
\description{
Checks if all required dependencies for the GUI are available.
}
\examples{
\dontrun{
checkGUIDependencies()
}

}
