% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-artifact.R
\name{cleanData}
\alias{cleanData}
\title{Clean data using an artifact removal pipeline}
\usage{
cleanData(
  x,
  steps = c("bad_channels", "ica"),
  bad_channel_method = "zscore",
  bad_channel_threshold = NULL,
  interpolation_method = "average",
  ica_method = "fastica",
  ica_classify_method = "kurtosis",
  output_assay = "cleaned"
)
}
\arguments{
\item{x}{A PhysioExperiment object.}

\item{steps}{Character vector of steps to run. Options: "bad_channels", "ica".}

\item{bad_channel_method}{Method for \code{\link{detectBadChannels}}.}

\item{bad_channel_threshold}{Threshold for bad channel detection.}

\item{interpolation_method}{Method for \code{\link{interpolateBadChannels}}.}

\item{ica_method}{Method for \code{\link{icaDecompose}}.}

\item{ica_classify_method}{Method for \code{\link{classifyICAComponents}}.}

\item{output_assay}{Name for the output assay.}
}
\value{
Modified PhysioExperiment with cleaned data.
}
\description{
Runs a configurable sequence of artifact removal steps.
}
