% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioCohort-class.R
\name{cohortData}
\alias{cohortData}
\alias{cohortData,PhysioCohort-method}
\alias{cohortData<-}
\alias{cohortData<-,PhysioCohort-method}
\title{Subject-level metadata table of a PhysioCohort}
\usage{
cohortData(x)

\S4method{cohortData}{PhysioCohort}(x)

cohortData(x) <- value

\S4method{cohortData}{PhysioCohort}(x) <- value
}
\arguments{
\item{x}{A \code{PhysioCohort}.}

\item{value}{A subject-level \code{DataFrame} (setter); one row per subject.}
}
\value{
\code{cohortData()} the subject-level \code{DataFrame}.
}
\description{
Subject-level metadata table of a PhysioCohort
}
