% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-resample.R
\name{decimate}
\alias{decimate}
\title{Decimate signal}
\usage{
decimate(x, factor, filter_order = 8L, output_assay = "decimated")
}
\arguments{
\item{x}{A PhysioExperiment object.}

\item{factor}{Integer decimation factor.}

\item{filter_order}{Order of the anti-aliasing lowpass filter.}

\item{output_assay}{Name for the output assay.}
}
\value{
A new PhysioExperiment with decimated data.
}
\description{
Downsamples by an integer factor with anti-aliasing filter.
}
