% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioLongitudinal-class.R
\name{design,PhysioLongitudinal-method}
\alias{design,PhysioLongitudinal-method}
\alias{design<-,PhysioLongitudinal-method}
\title{The design schema of a PhysioLongitudinal}
\usage{
\S4method{design}{PhysioLongitudinal}(object, ...)

\S4method{design}{PhysioLongitudinal}(object, ...) <- value
}
\arguments{
\item{object}{A \code{PhysioLongitudinal}.}

\item{...}{Ignored.}

\item{value}{A design \code{DataFrame} (setter).}
}
\value{
\code{design()} the design \code{DataFrame}.
}
\description{
A method for the \pkg{BiocGenerics} \code{design} generic, so it composes
cleanly with the rest of the Bioconductor ecosystem.
}
