% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-epoch.R
\name{epochTimes}
\alias{epochTimes}
\title{Get epoch time vector}
\usage{
epochTimes(x)
}
\arguments{
\item{x}{An epoched PhysioExperiment object.}
}
\value{
Numeric vector of times in seconds.
}
\description{
Returns the time vector for epoched data relative to event onset.
}
