% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-query.R
\docType{class}
\name{EventQuery-class}
\alias{EventQuery-class}
\title{EventQuery class for composable event filtering}
\description{
EventQuery class for composable event filtering
}
\section{Slots}{

\describe{
\item{\code{source}}{PhysioExperiment object}

\item{\code{filters}}{List of filter functions to apply}

\item{\code{resolved}}{Cached resolved events (NULL if not yet resolved)}
}}

