% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-query.R
\name{eventQuery}
\alias{eventQuery}
\title{Create an EventQuery from a PhysioExperiment}
\usage{
eventQuery(x)
}
\arguments{
\item{x}{A PhysioExperiment object}
}
\value{
An EventQuery object
}
\description{
Create an EventQuery from a PhysioExperiment
}
\examples{
pe <- PhysioExperiment(assays = list(raw = matrix(rnorm(400), 100, 4)),
                       samplingRate = 100)
pe <- addEvents(pe, onset = c(1, 2, 3), type = "stimulus")
q <- eventQuery(pe)
}
