% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-fft.R
\name{fftSignals}
\alias{fftSignals}
\title{Fast Fourier transform helper}
\usage{
fftSignals(x)
}
\arguments{
\item{x}{A \code{PhysioExperiment} object.}
}
\value{
The modified object containing an FFT assay.
}
\description{
Computes the discrete Fourier transform along the time axis of the default
assay and stores the magnitude spectrum in a new assay named \code{"fft"}.
}
