% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-query.R
\name{filterType}
\alias{filterType}
\title{Filter events by type}
\usage{
filterType(q, types)
}
\arguments{
\item{q}{An EventQuery object}

\item{types}{Character vector of event types to keep}
}
\value{
Modified EventQuery
}
\description{
Filter events by type
}
\examples{
pe <- PhysioExperiment(assays = list(raw = matrix(rnorm(400), 100, 4)),
                       samplingRate = 100)
pe <- addEvents(pe, onset = c(1, 2, 3),
                type = c("stimulus", "response", "stimulus"))
q <- filterType(eventQuery(pe), "stimulus")
resolveQuery(q)
}
