% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-query.R
\name{filterValue}
\alias{filterValue}
\title{Filter events by value}
\usage{
filterValue(q, values)
}
\arguments{
\item{q}{An EventQuery object}

\item{values}{Character vector of event values to keep}
}
\value{
Modified EventQuery
}
\description{
Filter events by value
}
\examples{
pe <- PhysioExperiment(assays = list(raw = matrix(rnorm(400), 100, 4)),
                       samplingRate = 100)
pe <- addEvents(pe, onset = c(1, 2, 3), type = "stimulus",
                value = c("target", "distractor", "target"))
q <- filterValue(eventQuery(pe), "target")
resolveQuery(q)
}
