% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker.R
\name{format.PhysioBiomarker}
\alias{format.PhysioBiomarker}
\title{Format a PhysioBiomarker as a compact string}
\usage{
\method{format}{PhysioBiomarker}(x, ...)
}
\arguments{
\item{x}{A \link{PhysioBiomarker}.}

\item{...}{Ignored.}
}
\value{
A length-1 character string \code{"name = value unit [ci]"}.
}
\description{
Format a PhysioBiomarker as a compact string
}
