% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-epoch.R
\name{grandAverage}
\alias{grandAverage}
\title{Grand average across subjects/samples}
\usage{
grandAverage(...)
}
\arguments{
\item{...}{PhysioExperiment objects or a list of them.}
}
\value{
A PhysioExperiment object with grand averaged data.
}
\description{
Computes grand average across multiple PhysioExperiment objects.
}
