% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-network.R
\name{graphLaplacian}
\alias{graphLaplacian}
\title{Compute graph Laplacian}
\usage{
graphLaplacian(adjacency, normalized = FALSE)
}
\arguments{
\item{adjacency}{An adjacency matrix.}

\item{normalized}{If TRUE, returns normalized Laplacian.}
}
\value{
The Laplacian matrix.
}
\description{
Calculates the graph Laplacian matrix.
}
\examples{
adj <- matrix(c(0, 1, 1, 1, 0, 1, 1, 1, 0), 3, 3)
L <- graphLaplacian(adj)
}
