% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker.R
\name{is.PhysioBiomarker}
\alias{is.PhysioBiomarker}
\title{Test whether an object is a PhysioBiomarker}
\usage{
is.PhysioBiomarker(x)
}
\arguments{
\item{x}{Any object.}
}
\value{
\code{TRUE} if \code{x} is a \link{PhysioBiomarker}, otherwise \code{FALSE}.
}
\description{
Test whether an object is a PhysioBiomarker
}
\examples{
is.PhysioBiomarker(physioBiomarker(1, "x"))
is.PhysioBiomarker(42)
}
\seealso{
\code{\link[=physioBiomarker]{physioBiomarker()}}
}
