% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-rereference.R
\name{isAverageReferenced}
\alias{isAverageReferenced}
\title{Check if data is average referenced}
\usage{
isAverageReferenced(x)
}
\arguments{
\item{x}{A PhysioExperiment object.}
}
\value{
Logical indicating if data is average referenced.
}
\description{
Check if data is average referenced
}
\examples{
pe <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(100), nrow = 10, ncol = 10)),
  samplingRate = 100
)
pe <- rereference(pe, ref_type = "average")
isAverageReferenced(pe)  # TRUE
}
