% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-network.R
\name{localEfficiency}
\alias{localEfficiency}
\title{Compute local efficiency}
\usage{
localEfficiency(adjacency, weighted = FALSE)
}
\arguments{
\item{adjacency}{An adjacency matrix.}

\item{weighted}{If TRUE, uses weighted paths.}
}
\value{
A numeric vector of local efficiency values.
}
\description{
Calculates the local efficiency for each node.
}
\examples{
adj <- matrix(c(0, 1, 1, 1, 0, 1, 1, 1, 0), 3, 3)
localEfficiency(adj)
}
