% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{nEvents}
\alias{nEvents}
\title{Get number of events}
\usage{
nEvents(x)
}
\arguments{
\item{x}{A PhysioEvents object.}
}
\value{
Integer count of events.
}
\description{
Get number of events
}
\examples{
events <- PhysioEvents(onset = c(1, 2, 3), type = "stimulus")
nEvents(events)
}
\references{
Delorme A, Makeig S (2004). "EEGLAB: an open source toolbox for analysis
of single-trial EEG dynamics." \emph{Journal of Neuroscience Methods},
134(1), 9-21.
}
\seealso{
\code{\link{getEvents}} for retrieving events,
\code{\link{PhysioEvents}} for the event constructor,
\code{\link{addEvents}} for appending events
}
