% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-network.R
\name{pathLength}
\alias{pathLength}
\title{Compute shortest path length}
\usage{
pathLength(adjacency, weighted = FALSE, use_cpp = TRUE)
}
\arguments{
\item{adjacency}{An adjacency matrix.}

\item{weighted}{If TRUE, uses edge weights as distances.}
}
\value{
A matrix of shortest path lengths.
}
\description{
Calculates the shortest path length between all node pairs using Floyd-Warshall.
}
\examples{
adj <- matrix(c(0, 1, 0, 1, 0, 1, 0, 1, 0), 3, 3)
pathLength(adj)
}
