% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioLongitudinal-class.R
\name{PhysioLongitudinal-methods}
\alias{PhysioLongitudinal-methods}
\alias{length,PhysioLongitudinal-method}
\alias{[,PhysioLongitudinal,ANY,ANY,ANY-method}
\alias{show,PhysioLongitudinal-method}
\title{Display, size and subsetting for PhysioLongitudinal}
\usage{
\S4method{length}{PhysioLongitudinal}(x)

\S4method{[}{PhysioLongitudinal,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{show}{PhysioLongitudinal}(object)
}
\arguments{
\item{x}{A \code{PhysioLongitudinal}.}

\item{i}{Session index/name/logical (for \code{[}).}

\item{j, drop}{Ignored (present for \code{[} generic compatibility).}

\item{...}{Ignored.}

\item{object}{A \code{PhysioLongitudinal} (for \code{show}).}
}
\value{
\code{length()} the number of sessions; \code{[} a subset
\code{PhysioLongitudinal}; \code{show()} is called for its side effect.
}
\description{
Display, size and subsetting for PhysioLongitudinal
}
