% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-network.R
\name{plotNetworkMetrics}
\alias{plotNetworkMetrics}
\title{Plot network metrics}
\usage{
plotNetworkMetrics(
  adjacency,
  metrics = c("degree", "clustering", "betweenness"),
  node_names = NULL,
  title = "Network Metrics"
)
}
\arguments{
\item{adjacency}{An adjacency matrix.}

\item{metrics}{Vector of metrics to compute: "degree", "clustering",
"betweenness", "eigenvector", "local_efficiency".}

\item{node_names}{Optional vector of node names.}

\item{title}{Plot title.}
}
\value{
A ggplot object.
}
\description{
Creates a bar plot of network metrics for each node.
}
\examples{
adj <- matrix(c(0, 1, 1, 1, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0), 4, 4)
plotNetworkMetrics(adj, node_names = c("Fz", "Cz", "Pz", "Oz"))
}
