% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-multichannel.R
\name{plotPSD}
\alias{plotPSD}
\title{Plot power spectral density}
\usage{
plotPSD(x, channels = NULL, sample = 1L, log_scale = TRUE, freq_range = NULL)
}
\arguments{
\item{x}{A PhysioExperiment object.}

\item{channels}{Integer vector of channel indices. If NULL, plots all.}

\item{sample}{Integer index for the sample (for 3D data).}

\item{log_scale}{Logical. If TRUE, uses log scale for power.}

\item{freq_range}{Numeric vector of length 2 specifying frequency range.}
}
\value{
A ggplot object.
}
\description{
Generates a PSD plot for specified channels.
}
