% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trials.R
\name{recordings}
\alias{recordings}
\title{The recordings held by a trial set}
\usage{
recordings(x)
}
\arguments{
\item{x}{A \code{physio_trial_set}.}
}
\value{
The named list of \code{MultiPhysioExperiment} recordings.
}
\description{
The recordings held by a trial set
}
\seealso{
\code{\link{trialSet}}
}
