% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-query.R
\name{resolveQuery}
\alias{resolveQuery}
\title{Resolve an EventQuery to get filtered events}
\usage{
resolveQuery(q)
}
\arguments{
\item{q}{An EventQuery object}
}
\value{
Data frame of filtered events
}
\description{
Resolve an EventQuery to get filtered events
}
\examples{
pe <- PhysioExperiment(assays = list(raw = matrix(rnorm(400), 100, 4)),
                       samplingRate = 100)
pe <- addEvents(pe, onset = c(1, 2, 3), type = "stimulus")
resolveQuery(eventQuery(pe))
}
