% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis-theme.R
\name{scale_color_physio}
\alias{scale_color_physio}
\alias{scale_colour_physio}
\alias{scale_fill_physio}
\title{Colorblind-safe discrete colour / fill scales for ggplot2}
\usage{
scale_color_physio(...)

scale_colour_physio(...)

scale_fill_physio(...)
}
\arguments{
\item{...}{Passed to \code{ggplot2::discrete_scale()}.}
}
\value{
A \code{ggplot2} discrete scale using \code{\link{physioPalette}}.
}
\description{
Colorblind-safe discrete colour / fill scales for ggplot2
}
\examples{
if (requireNamespace("ggplot2", quietly = TRUE)) {
  p <- ggplot2::ggplot(
      iris,
      ggplot2::aes(Sepal.Length, Sepal.Width, color = Species)) +
    ggplot2::geom_point() + scale_color_physio() + theme_physio()
}
}
