% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioLongitudinal-class.R
\name{sessions}
\alias{sessions}
\alias{sessions,PhysioLongitudinal-method}
\alias{sessions<-}
\alias{sessions<-,PhysioLongitudinal-method}
\title{Access the sessions of a PhysioLongitudinal}
\usage{
sessions(x)

\S4method{sessions}{PhysioLongitudinal}(x)

sessions(x) <- value

\S4method{sessions}{PhysioLongitudinal}(x) <- value
}
\arguments{
\item{x}{A \code{PhysioLongitudinal}.}

\item{value}{A named list / SimpleList of sessions (setter).}
}
\value{
\code{sessions()} a \code{SimpleList}; setter returns the object.
}
\description{
Access the sessions of a PhysioLongitudinal
}
