% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{show,PhysioEvents-method}
\alias{show,PhysioEvents-method}
\title{Show method for PhysioEvents}
\usage{
\S4method{show}{PhysioEvents}(object)
}
\arguments{
\item{object}{A PhysioEvents object.}
}
\value{
Returns \code{NULL} invisibly; called for the side effect of printing
a compact summary of the events to the console.
}
\description{
Show method for PhysioEvents
}
\seealso{
\code{\link{PhysioEvents}} for the constructor,
\code{\link{nEvents}} for event count
}
