% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPhysioExperiment-class.R
\name{streamNames}
\alias{streamNames}
\title{Names of the streams}
\usage{
streamNames(x)
}
\arguments{
\item{x}{A \code{MultiPhysioExperiment}.}
}
\value{
Character vector of stream names.
}
\description{
Names of the streams
}
