% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPhysioExperiment-class.R
\name{streamRates}
\alias{streamRates}
\title{Per-stream sampling rates}
\usage{
streamRates(x)
}
\arguments{
\item{x}{A \code{MultiPhysioExperiment}.}
}
\value{
Named numeric of sampling rates (Hz), one per stream.
}
\description{
Per-stream sampling rates
}
