% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiPhysioExperiment-class.R
\name{streams}
\alias{streams}
\alias{streams,MultiPhysioExperiment-method}
\alias{streams<-}
\alias{streams<-,MultiPhysioExperiment-method}
\title{Access the streams of a MultiPhysioExperiment}
\usage{
streams(x)

\S4method{streams}{MultiPhysioExperiment}(x)

streams(x) <- value

\S4method{streams}{MultiPhysioExperiment}(x) <- value
}
\arguments{
\item{x}{A \code{MultiPhysioExperiment}.}

\item{value}{A named list / SimpleList of \code{PhysioExperiment} streams.}
}
\value{
\code{streams()} a \code{SimpleList}; setter returns the updated object.
}
\description{
Access the streams of a MultiPhysioExperiment
}
