% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioCohort-class.R
\name{subject}
\alias{subject}
\title{Retrieve a single subject's longitudinal record by id}
\usage{
subject(x, id)
}
\arguments{
\item{x}{A \code{PhysioCohort}.}

\item{id}{A subject id (name).}
}
\value{
The subject's \code{PhysioLongitudinal}.
}
\description{
Retrieve a single subject's longitudinal record by id
}
\seealso{
\code{\link{subjects}}
}
