% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioLongitudinal-class.R
\name{subjectData}
\alias{subjectData}
\alias{subjectData,PhysioLongitudinal-method}
\alias{subjectData<-}
\alias{subjectData<-,PhysioLongitudinal-method}
\title{Subject-level metadata}
\usage{
subjectData(x)

\S4method{subjectData}{PhysioLongitudinal}(x)

subjectData(x) <- value

\S4method{subjectData}{PhysioLongitudinal}(x) <- value
}
\arguments{
\item{x}{A \code{PhysioLongitudinal}.}

\item{value}{A one-row subject \code{DataFrame} (setter).}
}
\value{
\code{subjectData()} the subject \code{DataFrame}.
}
\description{
Subject-level metadata
}
