% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioCohort-class.R
\name{subjectIds}
\alias{subjectIds}
\alias{nSubjects}
\alias{groups}
\title{Subject ids, subject count and group labels of a PhysioCohort}
\usage{
subjectIds(x)

nSubjects(x)

groups(x)
}
\arguments{
\item{x}{A \code{PhysioCohort}.}
}
\value{
\code{subjectIds()} a character vector; \code{nSubjects()} an integer;
\code{groups()} the \code{colData$group} vector (or \code{NULL}).
}
\description{
Subject ids, subject count and group labels of a PhysioCohort
}
