% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioCohort-class.R
\name{subjects}
\alias{subjects}
\alias{subjects,PhysioCohort-method}
\alias{subjects<-}
\alias{subjects<-,PhysioCohort-method}
\title{Subjects of a PhysioCohort}
\usage{
subjects(x)

\S4method{subjects}{PhysioCohort}(x)

subjects(x) <- value

\S4method{subjects}{PhysioCohort}(x) <- value
}
\arguments{
\item{x}{A \code{PhysioCohort}.}

\item{value}{A named list / SimpleList of \code{PhysioLongitudinal} (setter).}
}
\value{
\code{subjects()} a \code{SimpleList} of \code{PhysioLongitudinal}.
}
\description{
Subjects of a PhysioCohort
}
