% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-resample.R
\name{assaySamplingRates}
\alias{assaySamplingRates}
\title{Get sampling rates for all assays}
\usage{
assaySamplingRates(x)
}
\arguments{
\item{x}{A PhysioExperiment object.}
}
\value{
Named numeric vector of sampling rates.
}
\description{
Returns sampling rates associated with each assay.
}
