% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{channelInfo<-}
\alias{channelInfo<-}
\title{Set channel information}
\usage{
channelInfo(x) <- value
}
\arguments{
\item{x}{A PhysioExperiment object.}

\item{value}{A DataFrame with channel information.}
}
\value{
Modified PhysioExperiment object.
}
\description{
Updates channel metadata.
Channel information is stored in colData (columns = channels).
}
\examples{
pe <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(400), nrow = 100, ncol = 4)),
  samplingRate = 100
)

# Set channel info
channelInfo(pe) <- S4Vectors::DataFrame(
  label = c("Fz", "Cz", "Pz", "Oz"),
  type = rep("EEG", 4)
)
}
