% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{channelNames<-}
\alias{channelNames<-}
\title{Set channel names/labels}
\usage{
channelNames(x) <- value
}
\arguments{
\item{x}{A PhysioExperiment object.}

\item{value}{Character vector of channel names.}
}
\value{
Modified PhysioExperiment object.
}
\description{
Set channel names/labels
}
\examples{
pe <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(400), nrow = 100, ncol = 4)),
  samplingRate = 100
)
channelNames(pe) <- c("Fz", "Cz", "Pz", "Oz")
channelNames(pe)
}
