% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{channelNames}
\alias{channelNames}
\title{Get channel names/labels}
\usage{
channelNames(x)
}
\arguments{
\item{x}{A PhysioExperiment object.}
}
\value{
Character vector of channel names.
}
\description{
Get channel names/labels
}
\examples{
pe <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(400), nrow = 100, ncol = 4)),
  colData = S4Vectors::DataFrame(label = c("Fz", "Cz", "Pz", "Oz")),
  samplingRate = 100
)
channelNames(pe)  # c("Fz", "Cz", "Pz", "Oz")
}
