% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioExperiment-accessors.R
\name{defaultAssay}
\alias{defaultAssay}
\title{Retrieve the default assay name}
\usage{
defaultAssay(x)
}
\arguments{
\item{x}{A \code{PhysioExperiment} instance.}
}
\value{
Character scalar naming the first assay or \code{NA_character_} when absent.
}
\description{
Retrieve the default assay name
}
