% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioExperiment-methods.R
\name{dim,PhysioExperiment-method}
\alias{dim,PhysioExperiment-method}
\title{Dim method for PhysioExperiment}
\usage{
\S4method{dim}{PhysioExperiment}(x)
}
\arguments{
\item{x}{A PhysioExperiment object.}
}
\value{
Integer vector of dimensions.
}
\description{
Returns dimensions of the default assay.
}
\examples{
pe <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(400), nrow = 100, ncol = 4)),
  samplingRate = 100
)
dim(pe)  # 100 4
}
