% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioExperiment-methods.R
\name{duration}
\alias{duration}
\title{Get signal duration}
\usage{
duration(x)
}
\arguments{
\item{x}{A PhysioExperiment object.}
}
\value{
Duration in seconds.
}
\description{
Get signal duration
}
\examples{
pe <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(1000), nrow = 1000, ncol = 4)),
  samplingRate = 100
)
duration(pe)  # 10 seconds
}
