% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-query.R
\name{filterType}
\alias{filterType}
\title{Filter events by type}
\usage{
filterType(q, types)
}
\arguments{
\item{q}{An EventQuery object}

\item{types}{Character vector of event types to keep}
}
\value{
Modified EventQuery
}
\description{
Filter events by type
}
