% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-query.R
\name{filterValue}
\alias{filterValue}
\title{Filter events by value}
\usage{
filterValue(q, values)
}
\arguments{
\item{q}{An EventQuery object}

\item{values}{Character vector of event values to keep}
}
\value{
Modified EventQuery
}
\description{
Filter events by value
}
