% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{getElectrodePositions}
\alias{getElectrodePositions}
\title{Get electrode positions}
\usage{
getElectrodePositions(x)
}
\arguments{
\item{x}{A PhysioExperiment object.}
}
\value{
A data.frame with x, y, z columns or NULL if not set.
}
\description{
Get electrode positions
}
\examples{
pe <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(300), nrow = 100, ncol = 3)),
  colData = S4Vectors::DataFrame(label = c("Fz", "Cz", "Pz")),
  samplingRate = 100
)
pe <- applyMontage(pe, "10-20")
getElectrodePositions(pe)
}
