% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{getReference}
\alias{getReference}
\title{Get reference electrode}
\usage{
getReference(x)
}
\arguments{
\item{x}{A PhysioExperiment object.}
}
\value{
Character string of reference electrode or NULL.
}
\description{
Get reference electrode
}
\examples{
pe <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(400), nrow = 100, ncol = 4)),
  samplingRate = 100
)
pe <- setReference(pe, "Cz")
getReference(pe)  # "Cz"
}
