% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioExperiment-methods.R
\name{length,PhysioExperiment-method}
\alias{length,PhysioExperiment-method}
\title{Length method for PhysioExperiment}
\usage{
\S4method{length}{PhysioExperiment}(x)
}
\arguments{
\item{x}{A PhysioExperiment object.}
}
\value{
Integer number of time points.
}
\description{
Returns the number of time points.
}
\examples{
pe <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(400), nrow = 100, ncol = 4)),
  samplingRate = 100
)
length(pe)  # 100
}
