% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{nChannels}
\alias{nChannels}
\title{Get number of channels}
\usage{
nChannels(x)
}
\arguments{
\item{x}{A PhysioExperiment object.}
}
\value{
Integer number of channels.
}
\description{
Get number of channels
}
\examples{
pe <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(400), nrow = 100, ncol = 4)),
  samplingRate = 100
)
nChannels(pe)  # 4
}
