% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioExperiment-methods.R
\name{rbindPhysio}
\alias{rbindPhysio}
\title{Combine PhysioExperiment objects by time}
\usage{
rbindPhysio(x, y)
}
\arguments{
\item{x}{A PhysioExperiment object.}

\item{y}{A PhysioExperiment object to concatenate.}
}
\value{
Combined PhysioExperiment object.
}
\description{
Concatenates two PhysioExperiment objects along the time axis.
}
\examples{
pe1 <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(400), nrow = 100, ncol = 4)),
  samplingRate = 100
)
pe2 <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(400), nrow = 100, ncol = 4)),
  samplingRate = 100
)

# Concatenate in time
pe_concat <- rbindPhysio(pe1, pe2)
length(pe_concat)  # 200
}
