% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{removeEvents}
\alias{removeEvents}
\title{Remove events from a PhysioExperiment object}
\usage{
removeEvents(x, type = NULL, indices = NULL)
}
\arguments{
\item{x}{A PhysioExperiment object.}

\item{type}{Optional event types to remove. If NULL, removes all events.}

\item{indices}{Optional integer indices of events to remove.}
}
\value{
The modified PhysioExperiment object.
}
\description{
Remove events from a PhysioExperiment object
}
