% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{renameChannels}
\alias{renameChannels}
\title{Rename channels}
\usage{
renameChannels(x, old_names, new_names)
}
\arguments{
\item{x}{A PhysioExperiment object.}

\item{old_names}{Character vector of current names.}

\item{new_names}{Character vector of new names.}
}
\value{
Modified PhysioExperiment object.
}
\description{
Rename channels
}
\examples{
pe <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(400), nrow = 100, ncol = 4)),
  colData = S4Vectors::DataFrame(label = c("Fz", "Cz", "Pz", "Oz")),
  samplingRate = 100
)

# Rename channels
pe <- renameChannels(pe, c("Fz", "Cz"), c("F3", "C3"))
channelNames(pe)
}
