% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-query.R
\name{resolveQuery}
\alias{resolveQuery}
\title{Resolve an EventQuery to get filtered events}
\usage{
resolveQuery(q)
}
\arguments{
\item{q}{An EventQuery object}
}
\value{
Data frame of filtered events
}
\description{
Resolve an EventQuery to get filtered events
}
