% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{samplesToTime}
\alias{samplesToTime}
\title{Convert sample indices to times}
\usage{
samplesToTime(x, samples)
}
\arguments{
\item{x}{A PhysioExperiment object.}

\item{samples}{Integer vector of sample indices.}
}
\value{
Numeric vector of times in seconds.
}
\description{
Convert sample indices to times
}
