% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops-resample.R
\name{setAssaySamplingRate}
\alias{setAssaySamplingRate}
\title{Set sampling rate for a specific assay}
\usage{
setAssaySamplingRate(x, assay_name, rate)
}
\arguments{
\item{x}{A PhysioExperiment object.}

\item{assay_name}{Name of the assay.}

\item{rate}{Sampling rate for the assay.}
}
\value{
Modified PhysioExperiment object.
}
\description{
Set sampling rate for a specific assay
}
