% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{setChannelUnits}
\alias{setChannelUnits}
\title{Set channel units}
\usage{
setChannelUnits(x, units)
}
\arguments{
\item{x}{A PhysioExperiment object.}

\item{units}{Character vector or named list of units.}
}
\value{
Modified PhysioExperiment object.
}
\description{
Assigns physical units to channels.
}
\examples{
pe <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(400), nrow = 100, ncol = 4)),
  samplingRate = 100
)

# Set all channels to same unit
pe <- setChannelUnits(pe, "uV")
}
