% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{show,PhysioEvents-method}
\alias{show,PhysioEvents-method}
\title{Show method for PhysioEvents}
\usage{
\S4method{show}{PhysioEvents}(object)
}
\arguments{
\item{object}{A PhysioEvents object.}
}
\description{
Show method for PhysioEvents
}
