% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysioExperiment-methods.R
\name{show,PhysioExperiment-method}
\alias{show,PhysioExperiment-method}
\title{Show method for PhysioExperiment}
\usage{
\S4method{show}{PhysioExperiment}(object)
}
\arguments{
\item{object}{A PhysioExperiment object.}
}
\description{
Displays a summary of the PhysioExperiment object.
}
\examples{
pe <- PhysioExperiment(
  assays = list(raw = matrix(rnorm(400), nrow = 100, ncol = 4)),
  colData = S4Vectors::DataFrame(label = c("Fz", "Cz", "Pz", "Oz")),
  samplingRate = 100
)
pe  # Displays summary
}
