% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-signal.R
\name{timeIndex}
\alias{timeIndex}
\title{Time index helper}
\usage{
timeIndex(x)
}
\arguments{
\item{x}{A \code{PhysioExperiment} instance.}
}
\value{
Numeric vector of time points in seconds.
}
\description{
Computes a time vector for the default assay using the object's sampling
rate.
}
