% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events.R
\name{timeToSamples}
\alias{timeToSamples}
\title{Convert event times to sample indices}
\usage{
timeToSamples(x, times)
}
\arguments{
\item{x}{A PhysioExperiment object.}

\item{times}{Numeric vector of times in seconds.}
}
\value{
Integer vector of sample indices.
}
\description{
Convert event times to sample indices
}
